updated main.h file

+++ 0x07-pointers_arrays_strings/main.h
@@ -9,7 +9,7 @@
 #include <stddef.h>
 
 int _putchar(char c);
-void *_memset(void *s, int c, size_t n);
+void *_memset(void *s, int c, unsigned int n);
 void *_memcpy(void *dest, const void *src, size_t n);
 char *_strchr(char *s, char c);
 unsigned int _strspn(char *s, char *accept);
